### 2022-11-22 #4956 @editorialbot
[PRE REVIEW]: DistributedSparseGrids.jl: A Julia library implementing an Adaptive Sparse Grid collocation method
**Submitting author:** @baxmittens (<a href="http://orcid.org/0000-0001-9954-294X">maximilian bittens</a>)
**Repository:** https://github.com/baxmittens/DistributedSparseGrids.jl
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.3
