### 2022-11-22 #4956 @baxmittens
@danielskatz

For me the webpage https://bit.ly/joss-reviewers is unavailable. Is there any other source where I can find a list of possible reviewers?

Thank you in advance
