+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lib/lib/#Library","page":"Library","title":"Library","text":"","category":"section"},{"location":"lib/lib/#Contents","page":"Library","title":"Contents","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"Pages = [\"lib.md\"]\nDepth = 4","category":"page"},{"location":"lib/lib/#Functions","page":"Library","title":"Functions","text":"","category":"section"},{"location":"lib/lib/#Index","page":"Library","title":"Index","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"Pages = [\"lib.md\"]","category":"page"},{"location":"lib/lib/#Typedefs","page":"Library","title":"Typedefs","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"DistributedSparseGrids.AbstractSparseGrid\nDistributedSparseGrids.AbstractHierarchicalSparseGrid\nDistributedSparseGrids.PointDict","category":"page"},{"location":"lib/lib/#DistributedSparseGrids.AbstractSparseGrid","page":"Library","title":"DistributedSparseGrids.AbstractSparseGrid","text":"AbstractSparseGrid{N}\n\nAbstract Type\n\n`N` : Dimension of hierarchical sparse grid\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#DistributedSparseGrids.AbstractHierarchicalSparseGrid","page":"Library","title":"DistributedSparseGrids.AbstractHierarchicalSparseGrid","text":"AbstractHierarchicalSparseGrid{N,HCP}\n\nAbstract Type\n\n`N` : Dimension of hierarchical sparse grid\n`HCP<:AbstractHierarchicalCollocationPoint` : Collocation point type\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#DistributedSparseGrids.PointDict","page":"Library","title":"DistributedSparseGrids.PointDict","text":"PointDict{ N, HCP <: AbstractHierarchicalCollocationPoint{N}}\n\nTypedef for `Dict{SVector{N,Int},Dict{SVector{N,Int},HCP}}`\n\n`N` : Dimension of hierarchical sparse grid\n`HCP<:AbstractHierarchicalCollocationPoint` : Collocation point type\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#Structs","page":"Library","title":"Structs","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"DistributedSparseGrids.AdaptiveHierarchicalSparseGrid","category":"page"},{"location":"lib/lib/#DistributedSparseGrids.AdaptiveHierarchicalSparseGrid","page":"Library","title":"DistributedSparseGrids.AdaptiveHierarchicalSparseGrid","text":"AdaptiveHierarchicalSparseGrid{N,HCP}\n\nContainer for hierarchical collocation points \n\n# Fields\n\n`cpts::Vector{PointDict{N,HCP}}` : [`DistributedSparseGrids.PointDict`](@ref) with collocation points\n`pointSetProperties::SVector{N,Int}` : Pointset properties\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#General-functions","page":"Library","title":"General functions","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"init\ninterpolate\nintegrate\n","category":"page"},{"location":"lib/lib/#DistributedSparseGrids.init","page":"Library","title":"DistributedSparseGrids.init","text":"init(::Typ
[…]
