### 2022-11-22 #4956 @drvinceknight
> For me the webpage https://bit.ly/joss-reviewers is unavailable. Is there any other source where I can find a list of possible reviewers?

Hi @baxmittens, I've just checked and that page loads fine for me.
