+++ src/DistributedSparseGrids.jl
@@ -48,7 +48,8 @@ const PointDict{ N, HCP <: AbstractHierarchicalCollocationPoint{N}} = Dict{SVect
 	# Fields
 
 	`cpts::Vector{PointDict{N,HCP}}` : [`DistributedSparseGrids.PointDict`](@ref) with collocation points
-	`pointSetProperties::SVector{N,Int}` : Pointset properties
+	`pointSetProperties::SVector{N,Int}` : Pointset properties = [psp_1,...,psp_N], psp_i \in [1,2,3,4]. 
+		1=>`closed point set`, 2=>`open point set`, 3=>`left-open point set`, 4=>`right-open point set`
 """
 struct AdaptiveHierarchicalSparseGrid{N,HCP} <: AbstractHierarchicalSparseGrid{N,HCP}
 	cpts::Vector{PointDict{N,HCP}}
@@ -76,7 +77,7 @@ Initialize the sparse grid. Returns a `N`-dimensional sparse grid where only the
 
 # Constructor
 - `::Type{AHSG{N,HCP}}`: Define type of [`DistributedSparseGrids.ASHG`](@ref)
-- `cpts::Dict{SVector{N,Int},Dict{SVector{N,Int},HCP}}`: Dict cointaining all collocation points
+- `pointSetProperties::SVector{N,Int}`: Vector containing all pointset properties, see 
 
 # Example
 N = 1
