+++ src/DistributedSparseGrids.jl
@@ -85,7 +85,7 @@ Initialize the sparse grid. Returns a `N`-dimensional sparse grid where only the
 - `::Type{AHSG{N,HCP}}`: Define type of [`DistributedSparseGrids.ASHG`](@ref)
 - `pointSetProperties::SVector{N,Int}`: Vector containing all pointset properties.
 
-Pointset properties = [psp_1,...,psp_N], psp_i \in [1,2,3,4]. 
+Pointset properties = [psp_1,...,psp_N], psp_i in [1,2,3,4]. 
 1=>`closed point set`, 
 2=>`open point set`, 
 3=>`left-open point set`, 
