+++ docs/src/lib/lib.md
@@ -40,6 +40,7 @@ init_weights!
 init_weights_inplace_ops!
 distributed_init_weights!
 distributed_init_weights_inplace_ops!
+generate_next_level!
 ```
 
 ### Utils
