+++ src/AdaptiveSparseGrids/refinement.jl
@@ -83,7 +83,7 @@ function generate_next_level!(asg::AHSG{N,HCP}) where {N,HCP<:AbstractHierarchic
 end
 
 """
-    generate_next_level!(asg::AHSG{N,HCP}) where {N,HCP<:AbstractHierarchicalCollocationPoint{N}}
+    generate_next_level!(asg::AHSG{N,HCP}, tol::CT,maxlvl::Int) where {N,CT,CP<:AbstractCollocationPoint{N,CT},HCP<:AbstractHierarchicalCollocationPoint{N,CP}}
 
 Adaptively generate all collocation point of the next hierarchical level, where `norm(scaling_weight(cpt)) > tol && level(cpt)<maxlvl`.
 
