### 2022-11-25 #4956 @drvinceknight
@ericneiva @idoby @matt-graham and @diegomcarvalho I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “DistributedSparseGrids.jl: A Julia library implementing an Adaptive Sparse Grid collocation method ”. You can find more information at the top of this Github issue.
