### 2022-11-27 #4956 @diegomcarvalho
Hi, Vince ***@***.***),

I'm so sorry for this late reply. I was way from my mailbox. If it is still missing a reviewer, please, you can include me.

All the best, d.c

> On 25 Nov 2022, at 07:18, Vince Knight ***@***.***> wrote:
>
>
> @ericneiva <https://github.com/ericneiva> @idoby <https://github.com/idoby> @matt-graham <https://github.com/matt-graham> and @diegomcarvalho <https://github.com/diegomcarvalho> I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “DistributedSparseGrids.jl: A Julia library implementing an Adaptive Sparse Grid collocation method ”. You can find more information at the top of this Github issue.
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
