### 2022-11-28 #4956 @matt-graham
> .. I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

Hi @drvinceknight. I would be happy to review this if you are still looking for additional reviewers
