### 2022-12-09 #4956 @drvinceknight
> I don't want to rush you but is there anything speaking against starting the review process with "editorialbot start review"?

We aim to do this once enough reviewers are in place. I've been quite busy so thanks for your patience. Now we have two reviewers I'll do this.
