### 2022-12-09 #5003 @editorialbot
[REVIEW]: DistributedSparseGrids.jl: A Julia library implementing 1 an Adaptive Sparse Grid collocation method
**Submitting author:** @baxmittens (<a href="http://orcid.org/0000-0001-9954-294X">maximilian bittens</a>)
**Repository:** https://github.com/baxmittens/DistributedSparseGrids.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
