+++ paper.bib
@@ -17,7 +17,8 @@
   number={3},
   pages={1005--1034},
   year={2007},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi = {10.1137/100786356}
 }
 
 
@@ -40,7 +41,8 @@
   number={8},
   pages={3084--3113},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.jcp.2009.01.006}
 }
 
 @article{gates2015multilevel,
