### 2022-12-15 baxmittens/DistributedSparseGrids.jl#3 @ericneiva
joss-paper: Fix installation instructions and update release
* [ ] In the [install section of the readme](https://github.com/baxmittens/DistributedSparseGrids.jl#install),

```
import Pkg
Pkg.install("DistributedSparseGrids")
```

should read

```
import Pkg
Pkg.add("DistributedSparseGrids")
```

Moreover,

* [x] I would strongly recommend to update the current release version to `0.1.3`; I did not manage to pass tests on `0.1.1`.

See openjournals/joss-reviews#5003
