### 2022-12-15 baxmittens/DistributedSparseGrids.jl#4 @ericneiva
joss-paper: Elaborate on target audience
I think both the documentation and the paper would benefit from a clearer statement on who the target audience is. I think it suffices to name the main (potential) end users of this software package.

See openjournals/joss-reviews#5003
