### 2022-12-15 baxmittens/DistributedSparseGrids.jl#5 @ericneiva
joss-paper: Add contribution guidelines
@baxmittens, could you please add contribution guidelines for third-parties wishing to:

* Contribute to the software
* Report issues or problems with the software, or
* Seek support

When you add them, please verify they are linked [here](https://github.com/baxmittens/DistributedSparseGrids.jl/community).

See openjournals/joss-reviews#5003
