### 2022-12-15 baxmittens/DistributedSparseGrids.jl#6 @ericneiva
joss-paper: Performance benchmarks
I think the software package (and the paper) will enormously benefit from a performance benchmark sustaining the performance claims of the software. I think this will also significantly increase the interest in the software and its impact.

I would be very happy to see a plot "number of collocation points vs time", plotting the time required to generate increasingly refined grids for a given function (up to millions of collocation points).

But if you have a better idea for a performance benchmark, please go ahead.

See openjournals/joss-reviews#5003
