### 2022-12-15 #5003 @ericneiva
Hi, @baxmittens I have gone through my review and filed issues baxmittens/DistributedSparseGrids.jl#3, baxmittens/DistributedSparseGrids.jl#4, baxmittens/DistributedSparseGrids.jl#5, baxmittens/DistributedSparseGrids.jl#6 and baxmittens/DistributedSparseGrids.jl#7
