+++ src/DistributedSparseGrids.jl
@@ -185,7 +185,8 @@ function interpolate!(res::RT, asg::SG, x::VCT, stplvl::Int=numlevels(asg)) wher
 	fill!(res,0.0)
 	tmp = zero(res)
 	in_it = InterpolationIterator(asg,x,stplvl)
-	for cpt_set in in_it
+	#for cpt_set in in_it
+	for hcpt in in_it
 		#for hcpt in cpt_set
 		mul!(tmp,scaling_weight(hcpt),basis_fun(hcpt, x, 1))
 		add!(res,tmp)
