### 2023-01-09 baxmittens/DistributedSparseGrids.jl#6 @ericneiva
Hi, @baxmittens, I really appreciate the effort you devoted to this.

I agree that a performance benchmark is not easy to include, so maybe I could do with a more elaborate and precise justification of your performance claim.

I've carefully read your results and discussion. I wonder if what sets this package apart from others and sustain your performance claim are the (1) distributed function evaluation and the (2) usage of arbitrary return types.

When evaluating a function at all grid points, while other packages exploit (at most) shared-memory parallelism, you can exploit distributed-memory parallelism. As a result, your package is clearly more suitable, when a single evaluation is computation- and memory-intensive.

Also, in contrast to other packages, the user can choose the right (performant) return data type for the problem at hand, optimise allocations with inplace operators...

Did I get this right?
