### 2023-01-09 baxmittens/DistributedSparseGrids.jl#6 @baxmittens
@ericneiva
A bit off-topic, but I just read some of your papers. They are very interesting to me. I also work on a similar topic. If you want to, you can check out our paper [an hp-hierarchical framework for the finite element exterior calculus](https://arxiv.org/abs/2012.15581) where we extended [natural refinement](http://www.multires.caltech.edu/pubs/Krysl-2002-CHARMS.pdf) (adaptive refinement without hanging nodes based on the refinement of basis function rather than elements) to arbitrary dimensions for all simplicial elements of [douglas n. arnolds finite element zoo](https://fenicsproject.org/pub/workshops/fenics13/slides/Arnold.pdf).
