+++ paper.md
@@ -49,7 +49,7 @@ In the following, some key features of the implemented approach are listed.
 
 ### Arbitrary return types
 
-DistributedSparseGrids.jl defines an interface for ```HierarchicalCollocationPoint{N,CP,RT}``` where ```N``` is the number of dimensions, ```CP <: AbstractCollocationPoint{N,CT<:Real}```, and ```RT``` is an generic return type. ```RT``` can be conveniently defined as a type needed most suitable for studying the problem at hand. This can be a ```Float64```, a ````Vector{Float64}```` or a ```Matrix{Float64}```, for example. There also exists a project named [VTUFileHandler.jl](https://github.com/baxmittens/VTUFileHandler.jl) @bittens2022vtufilehandler implementing the operators needed to load complete finite element results in the vtu-format @schroeder2000visualizing into the sparse grid.
+[DistributedSparseGrids.jl](https://github.com/baxmittens/DistributedSparseGrids.jl) defines a ```HierarchicalCollocationPoint{N,CP,RT}``` where ```N``` is the number of dimensions, ```CP <: AbstractCollocationPoint{N,CT<:Real}```, and ```RT``` is a generic return type. ```RT``` can be conveniently defined as the type most suitable for studying the problem at hand, such as a ```Float64```, a ````Vector{Float64}```` or a ```Matrix{Float64}```, for example. Suppose the underlying physical problem stores its data in the VTU file format [@schroeder2000visualizing]. In that case, the Julia project [VTUFileHandler.jl](https://github.com/baxmittens/VTUFileHandler.jl) [@bittens2022vtufilehandler] can be used, which implements all operators needed to load complete result files into the sparse grid.
 
 - Nested one-dimensional Clenshaw-Curtis rule
 - Smolyak's sparse grid construction
