### 2023-01-12 baxmittens/DistributedSparseGrids.jl#5 @baxmittens
should be all fine now. I close this issue. if something is wrong or missing with my contribution guidelines, feel free to reopen.
