### 2023-01-12 baxmittens/DistributedSparseGrids.jl#6 @baxmittens
I added some subchapters to the paper. I close this for now and generate a paper proof in the review repo.
Reopen this if necessary.
