### 2023-01-12 #5003 @baxmittens
I altered the paper according to the discussion in [issue 6](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/6). I only installed julia 1.8.3 now and ran DistributedSparseGrid.jl on it. There are some issues I have to fix, some of which are mentioned in [issue 3](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/3) and [issue 7](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/7) but I think I found even more. I will need some time (approx one week) to fix this.
