+++ src/DistributedSparseGrids.jl
@@ -6,7 +6,6 @@ using Distributed
 using ProgressMeter
 
 include(joinpath(".","CollocationPoints.jl"))
-include(joinpath(".","AdaptiveSparseGrids","plotting.jl"))
 
 
 """
@@ -508,6 +507,7 @@ function integrate(wasg::SG,skipdims::Vector{Int}) where {N,RT,CT,CP<:AbstractCo
 	return asg
 end
 
+include(joinpath(".","AdaptiveSparseGrids","plotting.jl"))
 export CollocationPoint, HierarchicalCollocationPoint, AHSG, init, generate_next_level!, init_weights!, distributed_init_weights!, init_weights_inplace_ops!, distributed_init_weights_inplace_ops!, integrate, interpolate
 
 end # module
