+++ README.md
@@ -201,6 +201,12 @@ for i = 1:20
 cpts = generate_next_level!(asg, 1e-5, 20)
 init_weights!(asg, collect(cpts), fun1)
 end
+
+# plot
+import PlotlyJS
+surfplot = PlotlyJS.surface(asg, 100)
+gridplot = PlotlyJS.scatter3d(asg)
+PlotlyJS.plot([surfplot, gridplot])
 ```
 
 <img src="https://user-images.githubusercontent.com/100423479/193813765-0b7ce7b2-639a-48d3-831d-7bd5639c9fd3.PNG" width="400" height="400" />
