+++ paper.bib
@@ -30,7 +30,8 @@
   number={5},
   pages={2309--2345},
   year={2008},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi = {10.1137/060663660}
 }
 
 @article{ma2009adaptive,
@@ -49,7 +50,8 @@
   title={A multilevel adaptive sparse grid stochastic collocation approach to the non-smooth forward propagation of uncertainty in discretized problems},
   author={Gates, Robert L and Bittens, Maximilian R},
   journal={arXiv preprint arXiv:1509.01462},
-  year={2015}
+  year={2015},
+  doi = {10.48550/arXiv.1509.01462}
 }
 
 @article{bittens2022vtufilehandler,
@@ -71,5 +73,6 @@
   number={5},
   pages={20--27},
   year={2000},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/38.865875}
 }
