### 2023-01-17 baxmittens/DistributedSparseGrids.jl#4 @baxmittens
I would like to disagree here politely. There are at least three other implementations of sparse grids, which I cite to show the significance of sparse grid approximations. I explain the necessity of my package by distinguishing myself from the other packages.
The package has been written to the best of my abilities to be universally applicable. For example, if there is any functional relation y=f(x), this package can be used to sample that function.
