### 2023-01-17 baxmittens/DistributedSparseGrids.jl#3 @baxmittens
I think I altered the code appropriately. Reopen if necessary.
