### 2023-01-17 baxmittens/DistributedSparseGrids.jl#4 @ericneiva
Hi, @baxmittens,

I am not saying I did not get an idea of your target audience by reading the _whole_ statement of need section of your paper, I am merely suggesting a _clearer_ (shorter) statement of your target audience.

> The package has been written to the best of my abilities to be universally applicable. For example, if there is any functional relation y=f(x), this package can be used to sample that function.

I agree that anyone can use your package to sample f(x), but I think you designed it based on a scenario where evaluation of f(x) is time-consuming or memory-heavy. So, I would say your end users are, above all, people who want to evaluate that type of functions, right?

Thereby, my suggestion is to put that into a single sentence. Something like adding to the first sentence of the statement of need
> "[DistributedSparseGrids.jl](https://github.com/baxmittens/DistributedSparseGrids.jl) is a Julia package for integrating and interpolating functions with generic return types, _whose evaluation is time-consuming and/or memory-demanding._".

Anything of the sort would do for me.

But I leave this as a suggestion to improve your paper. I know it's a detail. So, I am closing the issue.
