+++ .github/workflows/CI.yml
@@ -11,7 +11,7 @@ jobs:
       matrix:
         version:
           - '1.7'
-          - 'nightly'
+          - '1.8.5'
         os:
           - ubuntu-latest
         arch:
