### 2023-02-07 baxmittens/DistributedSparseGrids.jl#8 @matt-graham
Purpose of additional files in test directory unclear
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5003

As far as I can tell from the GitHub Actions CI workflow logs, the only test run as part of the CI is the single test in `test/runtests.jl`. There are a series of other files in the `test` directory, many of which contain only function definitions, and in general the intent of the code in these files is difficult to understand due to the use of non-descriptive variable names, presence of lots of commented out code and lack of comments describing intent. Ideally the files should be either moved to another directory if they're not intended to be tests, refactored in to a additional test(set)s using the `Test` package as in the `runtests.jl` file or removed from the repository if they are old. no longer needed, files.
