### 2023-02-07 baxmittens/DistributedSparseGrids.jl#10 @matt-graham
Existing test in `test/runtests.jl` needs more explanation and expansion
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5003

My understanding is that the single test in `test/runtest.jl` ,  constructs an initial sparse grid, adaptive refines it based on the test function

$$
  f(x) = \frac{1-\exp\left(-\left|2 - (x_1-1)^2 - (x_2-1)^2\right| - 0.01\right)}{\left|2 - (x_1-1)^2 - (x_2-1)^2\right| + 0.01}
$$

computes an approximation to the integral

$$
  I = \int_{[-1,1]^2} f(x) \mathrm{d}x
$$

and then compares this approximation to $I$ to the reference value 2.33905.

Currently what is being tested is not that clear. There are numerous [magic numbers](https://en.wikipedia.org/wiki/Magic_number_(programming)) in the code which need to be better explained - for example why is the number of iterations in the adaptive refinement loop 27, why is the number of initial levels 4, why are `tol` and `maxlevels` arguments to `generate_next_level!` of 1e-4 and 30 used (note also the comment describing these parameters is out of date), why is a tolerance of 1e-3 in the final comparison to the reference value? The provenance of the reference value of 2.33905 is also unclear - is this an analytic result, or the output of another numerical integration method? It is also unclear if the value used is the exact decimal representation or a truncation - if the latter, why do you truncate to a precision of 1e-5 when a tolerance of 1e-3 is used in the comparison?

The use of the `Distributed` package to distribute the computation over multiple workers is also not very clearly explained, and it seems like having this level of complexity for the only test is likely to lead to a very difficult to debug test failure if the test starts failing.

Ideally it would be good to have a suite of tests doing similar comparisons for several different test functions (for example the functions already defined in `test/testfuncs.jl` labelled as from the _Integration Test Problems Toolbox_, Michael Baudin), varying initial grid sizes and number of adaptive refinement steps, and on both a single and multiple workers, with the reference values for the comparison and tolerances needing to be set appropriately for each test case.
