### 2023-02-07 baxmittens/DistributedSparseGrids.jl#11 @matt-graham
Increasing coverage of tests
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5003

As discussed in #10 current there is only a single test in the automated test suite in `test/runtests.jl` which checks whether the output of the `integrate` function on an adaptively refined sparse grid for a particular test function is close to a reference value.

While this is a useful test, ideally we would have tests covering more cases of this task (as detailed in #10) but also more granular unit tests of all the methods / types exposed by the API, and also similar integration tests checking the values computed by `interpolate` against reference values.
