### 2023-02-07 baxmittens/DistributedSparseGrids.jl#12 @matt-graham
Indentation of docstrings causing rendering as code blocks
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5003

The indentation in the docstrings in

https://github.com/baxmittens/DistributedSparseGrids.jl/blob/ceacb12a067aa1222ee93f8facbccc9b14238016/src/DistributedSparseGrids.jl#L11-L58

is causing the text to be rendered incorrectly as code blocks in the HTML output

https://baxmittens.github.io/DistributedSparseGrids.jl/dev/lib/lib/#Typedefs
