### 2023-02-07 baxmittens/DistributedSparseGrids.jl#13 @matt-graham
Better high level description of algorithms implemented
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5003

The description of what the algorithms implemented in this package are used for currently given in the introduction in the README

https://github.com/baxmittens/DistributedSparseGrids.jl/blob/ceacb12a067aa1222ee93f8facbccc9b14238016/README.md?plain=1#L12-L14

is both a bit vague and also likely to be difficult for non-specialists to understand given the use of a lot of non-explained jargon (tensor-product discretizations, Smolyak's quadrature rule, adaptive sparse grid collocation, local hiearchical Lagrangian basis). Ideally each of these terms should either be explained within the text or at the least an appropriate reference linked to allow readers to look up the meanings themselves. It would also be good to give a succinct but concrete high level description of what numerical tasks the algorithms in the package are intended to address - that is numerical interpolation or integration of functions, which for example defines what classes of functions can be handled (as far as  I can tell scalar valued functions on hyperrectangular domains, though its unclear if there is any built in support for functions on domains other than $[-1,1]^N$).
