### 2023-02-07 #5003 @matt-graham
Hi @baxmittens.

Apologies for the delay in getting on to this and responding, I'm currently away on leave but will be back next week. I've added and linked some issues describing some of the comments and requests for changes I have so far.

As an additional question with regards to the paper - could you clarify what the role of the second author, Roberts L. Gates, was in the development of the package and writing of the paper? As far as I can see they have not made any commits on the GitHub repository.
