### 2023-02-07 #5003 @baxmittens
Hi @matt-graham

all good.
Thank you for your comments. I will work on fixing these issues until next week.

Initially, this code was a collaboration work of robert and me in 2015 and I think we worked on the code in equal parts.
As a result, we published a [arxiv paper](https://arxiv.org/abs/1509.01462)
There exists different (private) repos with different versions of this code (this is a bit of a mess which is my fault).
I had to move it to our internal GitLab first to be able to use the code at my work place and then had to move it back to GitLab to make it available to external project partners.

I wrote the joss paper...but credit where credit is due. I think this will be fine, right?
