+++ src/DistributedSparseGrids.jl
@@ -449,6 +449,19 @@ function integrate(asg::SG) where {N,CP,RT,HCP<:AbstractHierarchicalCollocationP
 	return res
 end
 
+function integrate_inplace_ops(asg::SG) where {N,CP,RT,HCP<:AbstractHierarchicalCollocationPoint{N,CP,RT}, SG<:AbstractHierarchicalSparseGrid{N,HCP}}
+	println("N=$N,CP=$CP,RT=$RT")
+	res = zero(first(asg).scaling_weight)
+	tmp = zero(first(asg).scaling_weight)
+	for cpt in asg
+		bf = integral_basis_fun(cpt)
+		sw = scaling_weight(cpt)
+		mul!(tmp, sw, bf)
+		add!(res, tmp) 
+	end
+	return res
+end
+
 #function integrate(asg::SG,fun::Function) where {N,CP,RT,HCP<:AbstractHierarchicalCollocationPoint{N,CP,RT}, SG<:AbstractHierarchicalSparseGrid{N,HCP}}
 #	println("N=$N,CP=$CP,RT=$RT")
 #	maxp = maxporder(asg)
