+++ test/runtests.jl
@@ -21,16 +21,110 @@ function sparse_grid(N::Int,pointprobs,nlevel=6,RT=Float64,CT=Float64)
 end
 
 @testset "DistributedSparseGrids.jl" begin
+
+# Define normal distribution
+norml(x,μ,σ) = 1/(σ*sqrt(2*pi))*exp(-0.5*((x[1]-μ)/σ)^2)
+
+# Define 3 1d sparse grids with different points sets
+asg11 = sparse_grid(1, @SVector [1]) 
+asg12 = sparse_grid(1, @SVector [2]) 
+asg13 = sparse_grid(1, @SVector [3]) 
+
+# 1d Normal distribution with μ=0.0 and σ=0.1
+norml1d(x::SVector{1,Float64},ID::String) = norml(x[1],0.0,0.1)
+
+# Init sparse-grids
+init_weights!(asg11, norml1d)
+init_weights!(asg12, norml1d)
+init_weights!(asg13, norml1d)
+
+# adaptive refine
+while true
+  # call generate_next_level! with tol=1e-4 and maxlevels=30
+  cpts11 = generate_next_level!(asg11, 1e-4, 30)
+  cpts12 = generate_next_level!(asg12, 1e-4, 30)
+  cpts13 = generate_next_level!(asg13, 1e-4, 30)
+  #
+  if isempty(cpts11) && isempty(cpts12) && isempty(cpts13)
+    break
+  end
+  # calculate weights on all worker
+  init_weights!(asg11, collect(cpts11), norml1d)
+  init_weights!(asg12, collect(cpts12), norml1d)
+  init_weights!(asg13, collect(cpts13), norml1d)
+end
+
+# integrating over pdf should always result in 1.0
+@test  isapprox(1.0, integrate(asg11), atol=1e-4)
+@test  isapprox(1.0, integrate(asg12), atol=1e-4)
+@test  isapprox(1.0, integrate(asg13), atol=1e-4)
+
+# Define 2 2d sparse grids with different points sets
+asg21 = sparse_grid(2, @SVector [1,1]) 
+asg22 = sparse_grid(2, @SVector [2,2]) 
+
+# 2d Normal distribution with μ=0.0 and σ=0.1
+norml2d(x::SVector{2,Float64},ID::String) = norml(x[1],0.0,0.1)*norml(x[2],0.0,0.1)
+
+# init weights
+init_weights!(asg21, norml2d)
+init_weights!(asg22, norml2d)
+
+# adaptive refine
+while true
+  # call generate_next_level! with tol=1e-4 and maxlevels=30
+  cpts21 = generate_next_level!(asg21, 1e-4, 30)
+  cpts22 = generate_next_level!(asg22, 1e-4, 30)
+  #
+  if isempty(cpts21) && isempty(cpts22)
+    break
+  end
+  # calculate weights on all worker
+  init_weights!(asg21, collect(cpts21), norml2d)
+  init_weights!(asg22, collect(cpts22), norml2d)
+end
+
+# integrating over pdf should always result in 1.0
+@test  isapprox(1.0, integrate(asg21), atol=1e-4)
+@test  isapprox(1.0, integrate(asg22), atol=1e-4)
+
+# Define 1 3d sparse grids
+asg31 = sparse_grid(3, @SVector [1,1,1]) 
+
+# Normal distribution with μ=0.0 and σ=0.1
+norml3d(x::SVector{3,Float64},ID::String) = norml(x[1],0.0,0.1)*norml(x[2],0.0,0.1)*norml(x[3],0.0,0.1)
+
+init_weights!(asg31, norml3d)
+
+# adaptive refine
+while true
+  # call generate_next_level! with tol=1e-4 and maxlevels=30
+  cpts31 = generate_next_level!(asg31, 1e-4, 30)
[…]
