+++ README.md
@@ -11,6 +11,15 @@ For an alternative implementation, see [AdaptiveSparseGrids.jl](https://github.c
 
 ## Introduction
 
+Any integral 
+
+$$
+  \int_{[a_1,b_1]}\cdot\cdot\cdot\int_{[a_n,b_n]} f(x_1,...,x_n) \mathrm{d}x_1\cdot\cdot\cdot\mathrm{d}x_n
+$$
+
+can be mapped onto the hypercube $[-1,1]^n$ by means of coordinate transformation.
+
+
 To mitigate the "curse of dimensionality" that occurs in the integration or interpolation of high-dimensional functions using tensor-product discretizations, sparse grids use Smolyak's quadrature rule. This is particularly useful if the evaluation of the underlying function is costly. In this library, an Adaptive Sparse Grid Collocation method with a local hierarchical Lagrangian basis, first proposed by [Ma and Zabaras (2010)](https://www.sciencedirect.com/science/article/pii/S002199910900028X), is implemented. For more information about the construction of Sparse Grids, see e.g. [Gates and Bittens (2015)](https://arxiv.org/abs/1509.01462).
 
 ## Install
@@ -34,6 +43,11 @@ Pkg.add("DistributedSparseGrids")
 -	experimental: integration over $X_{\sim (i)}$ (the $X_{\sim (i)}$  notation indicates the set of all variables except $X_{i}$).
 
 ## Usage
+
+### General remarks
+
+-	The quality of the error prediction depends on the number of collocation points. Therefore, for only a few collocation points, adaptive refinement may fail. Therefore it is recommended to generate some initial levels before using adaptive refinement (see examples).
+-	The interpolation is based on a local Lagrangian basis. Functions with discontinuities cannot be approximated.
 
 ### Point sets
 

[…]
