### 2023-02-13 baxmittens/DistributedSparseGrids.jl#13 @baxmittens
altered the description and added a bit of context.
