### 2023-02-13 baxmittens/DistributedSparseGrids.jl#11 @baxmittens
I've added some more tests.

Testing only integration is fine, though. For the initialization of weights, interpolation is called since for the calculation of the weights the interpolator of the underlying level is used.
