### 2023-02-13 baxmittens/DistributedSparseGrids.jl#10 @baxmittens
I've added some context for the `magic numbers` and added some additional tests (integration of a normal pdf in different dimensions).

- The function $f(x)$ is chosen because it is the most complex example of what the grid can integrate.
- The value 2.33905 is the analytical result, I have added a link to wolframalpha, where you can see the calculation.
- I replaced the floor loop (magic number 27) with a while loop
- 1e-4 is chosen by knowledge since for 1e-5 one would need over 200.000 additional collocation points which would take some time which would be annoying if the tests are run and would add little to nothing for testing the actual function.

surely this is an engineered example but not a bad one. this is applied math after all.

i hope with the additional examples and the added context, this will be fine
