### 2023-02-13 baxmittens/DistributedSparseGrids.jl#8 @baxmittens
deleted unnecessary files
