### 2023-02-13 baxmittens/DistributedSparseGrids.jl#10 @baxmittens
the functions in test/testfuncs.jl were mainly examples of functions that the sparse grid cannot integrate. I deleted the file and added an example of integration of normal pdf, which is a common and difficult example
