+++ Project.toml
@@ -12,6 +12,7 @@ Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 FastGaussQuadrature = "442a2c76-b920-505d-bb47-c5924d526838"
+AltInplaceOpsInterface = "d59240d3-86bd-48c4-8b44-fa29ea3b0621"
 
 
 [compat]
