+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lib/lib/#Library","page":"Library","title":"Library","text":"","category":"section"},{"location":"lib/lib/#Contents","page":"Library","title":"Contents","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"Pages = [\"lib.md\"]\nDepth = 4","category":"page"},{"location":"lib/lib/#Functions","page":"Library","title":"Functions","text":"","category":"section"},{"location":"lib/lib/#Index","page":"Library","title":"Index","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"Pages = [\"lib.md\"]","category":"page"},{"location":"lib/lib/#Typedefs","page":"Library","title":"Typedefs","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"DistributedSparseGrids.AbstractSparseGrid\nDistributedSparseGrids.AbstractHierarchicalSparseGrid\nDistributedSparseGrids.PointDict","category":"page"},{"location":"lib/lib/#DistributedSparseGrids.AbstractSparseGrid","page":"Library","title":"DistributedSparseGrids.AbstractSparseGrid","text":"AbstractSparseGrid{N}\n\nAbstract Type\n\nN : Dimension of hierarchical sparse grid\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#DistributedSparseGrids.AbstractHierarchicalSparseGrid","page":"Library","title":"DistributedSparseGrids.AbstractHierarchicalSparseGrid","text":"AbstractHierarchicalSparseGrid{N,HCP}\n\nAbstract Type\n\nN : Dimension of hierarchical sparse grid HCP<:AbstractHierarchicalCollocationPoint : Collocation point type\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#DistributedSparseGrids.PointDict","page":"Library","title":"DistributedSparseGrids.PointDict","text":"PointDict{ N, HCP <: AbstractHierarchicalCollocationPoint{N}}\n\nTypedef for Dict{SVector{N,Int},Dict{SVector{N,Int},HCP}}\n\nN : Dimension of hierarchical sparse grid HCP<:AbstractHierarchicalCollocationPoint : Collocation point type\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#Structs","page":"Library","title":"Structs","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"DistributedSparseGrids.CollocationPoint\nDistributedSparseGrids.HierarchicalCollocationPoint\nDistributedSparseGrids.AdaptiveHierarchicalSparseGrid","category":"page"},{"location":"lib/lib/#DistributedSparseGrids.CollocationPoint","page":"Library","title":"DistributedSparseGrids.CollocationPoint","text":"CollocationPoint{N,CT} <: AbstractCollocationPoint{N,CT}\n\nA collocation point.\n\nFields\n\ni_multi::SVector{N,Int} : The i-th item of this Vector represents to the level of the collocation point in the i-th dimension. pt_idx::SVector{N,Int} : The i-th item of this Vector represents to the point index of the collocation point in the i-th dimension. coords::SVector{N,CT} : Coordinates of the collocation point. interv::SVector{N,SVector{2,CT}} : The i-th item of the Vector defines the non-zero interval of the associated basis function in the i-th dimension.\n\n\n\n\n\n","category":"type"},{"location":"lib/lib/#DistributedSparseGrids.HierarchicalCollocationPoint","page":"Librar
[…]
