### 2023-02-20 baxmittens/DistributedSparseGrids.jl#12 @baxmittens
altered docstrings
