### 2023-02-20 #5003 @baxmittens
Hi @matt-graham

- I [altered](https://github.com/baxmittens/DistributedSparseGrids.jl/blob/main/README.md#Introduction) the description of the implemented algorithm according to [issue 13](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/13)
- [fixed](https://baxmittens.github.io/DistributedSparseGrids.jl/dev/lib/lib/) the indentation according to [issue 12](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/12)
- [Extended test coverage and added context](https://github.com/baxmittens/DistributedSparseGrids.jl/blob/main/test/runtests.jl) according to [issue 11](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/11) and [issue 10](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/10)
- deleted unnecessary files in ".test/" according to [issue 8](https://github.com/baxmittens/DistributedSparseGrids.jl/issues/8)

could you please check if your requests are satisfied?

Thx Max
