### 2023-02-21 baxmittens/DistributedSparseGrids.jl#11 @matt-graham
Okay that makes sese with regards to interpolation versus integration. Thanks for adding the additional tests
