### 2023-03-06 #5003 @drvinceknight
@baxmittens can you change the title of https://zenodo.org/record/7673697#.ZAW4PS-l1pQ to be the title of this paper `DistributedSparseGrids.jl: A Julia library implementing 1 an Adaptive Sparse Grid collocation method ` (if the DOI is updated please confirm it here)
