### 2023-03-06 #5003 @baxmittens
The "1" is a copy-paste mistake.
I would change it to: `DistributedSparseGrids.jl: A Julia library implementing an Adaptive Sparse Grid collocation method`
if that is ok
