### 2023-03-06 #5003 @drvinceknight
Please note there is a typo in the title of the paper on this issue: `DistributedSparseGrids.jl: A Julia library implementing 1 an Adaptive Sparse Grid collocation method` should be `DistributedSparseGrids.jl: A Julia library implementing an Adaptive Sparse Grid collocation method`.
