### 2023-03-06 #5003 @baxmittens
@drvinceknight
for me the proof looks good.
