### 2023-03-07 #5003 @danielskatz
@baxmittens - I've suggested some small changes in https://github.com/baxmittens/DistributedSparseGrids.jl/pull/14

In addition, please add a country to your affiliation.
