### 2023-03-07 #5003 @danielskatz
For your co-author, JOSS practice is to use the affiliation "Independent Researcher", and again to add a country.
