### 2023-03-07 #5003 @danielskatz
this is likely a DNS caching issue - perhaps you didn't wait a few minutes before trying the DOI?  In any case, it works fine for me, and should eventually for you too.
![Uploading Screenshot 2023-03-07 at 8.59.42 AM.png…]()
